Don't use http_prefix.
It *used* to work how I wanted it to work (only for assets, not .html links), but someone complained
and they changed it. They should probably have a separate cdn_prefix setting or something, but I
discovered that I can indeed host my entire site directly off of cloudfront afterall, so I'm going
to give that a whirl instead of trying to make this work how I want.