Publishing this article a bit early.
I've seen a couple of mentions to Ember authentication this week, with guides probably better than
this one, so this may already be too late. Might as well get it up with a caveat that I wrote it
a month and a half ago, in case someone still has some use for it.